chore: amplify config moved on top of the element tree
